complete manual selection implemantation

+++ src/main/kotlin/com/travelsmartplus/routes/UserRoutes.kt
@@ -7,13 +7,17 @@ import com.travelsmartplus.models.responses.HttpResponses.BAD_REQUEST
 import com.travelsmartplus.models.responses.HttpResponses.FAILED_DELETE_USER
 import com.travelsmartplus.models.responses.HttpResponses.FAILED_EDIT_USER
 import com.travelsmartplus.models.responses.HttpResponses.FAILED_SETUP
+import com.travelsmartplus.models.responses.HttpResponses.FORBIDDEN
 import com.travelsmartplus.models.responses.HttpResponses.INTERNAL_SERVER_ERROR
 import com.travelsmartplus.models.responses.HttpResponses.NOT_FOUND
 import com.travelsmartplus.utils.HashingService
 import com.travelsmartplus.utils.Validator
 import io.ktor.http.*
 import io.ktor.server.application.*
+import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
 import io.ktor.server.plugins.*
+import io.ktor.server.plugins.ContentTransformationException
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -23,8 +27,10 @@ import io.ktor.server.routing.*
  * @author Gabriel Salas
  */
 
+val dao = UserDAOFacadeImpl()
+
 fun Route.userRoutes() {
-    val dao = UserDAOFacadeImpl()
+
     val hashingService = HashingService()
 
     route("/user/{id}") {
@@ -32,8 +38,16 @@ fun Route.userRoutes() {
         get {
             try {
                 val id = call.parameters["id"]?.toIntOrNull() ?: throw BadRequestException("Missing parameter")
-                val user = dao.getUser(id) ?: throw NotFoundException()
-                call.respond(HttpStatusCode.OK, user)
+                val requesterId = call.principal<JWTPrincipal>()!!.payload.getClaim("userId").asString().toInt()// User ID in JWT Token
+
+                // Check if user is owner or admin belongs to the same org
+                if (requesterId == id || isAdminSameOrg(requesterId, id)) {
+                    val user = dao.getUser(id) ?: throw NotFoundException()
+                    call.respond(HttpStatusCode.OK, user)
+                } else {
+                    call.respond(HttpStatusCode.Forbidden, FORBIDDEN)
+                }
+
             } catch (e: BadRequestException) {
                 e.printStackTrace()
                 call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
@@ -47,15 +61,29 @@ fun Route.userRoutes() {
         }
 
         // Update user
-        post {
+        put {
             try {
                 val id = call.parameters["id"]?.toIntOrNull() ?: throw BadRequestException("Missing parameter")
+                val requesterId = call.principal<JWTPrincipal>()!!.payload.getClaim("userId").asString().toInt()// User ID in JWT Token
                 val editedUser = call.receive<User>()
-                val user = dao.editUser(id, editedUser)
-                call.respond(HttpStatusCode.Created, user)
+
+                // Check if user is owner or admin belongs to the same org
+                if (requesterId == id|| isAdminSameOrg(requesterId, id)) {
+                    val user = dao.editUser(id, editedUser)
+                    call.respond(HttpStatusCode.Created, user)
+                } else {
+                    call.respond(HttpStatusCode.Forbidden, FORBIDDEN)
+                }
+
             } catch (e: BadRequestException) {
                 e.printStackTrace()
                 call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
+            } catch (e: ContentTransformationException) {
+                e.printStackTrace()
+                call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
+            } catch (e: NotFoundException) {
+                e.printStackTrace()
+                call.respond(HttpStatusCode.NotFound, NOT_FOUND)
             } catch (e: Exception) {
                 e.printStackTrace()
                 call.respond(HttpStatusCode.InternalServerError, FAILED_EDIT_USER)
@@ -65,12 +93,23 @@ fun Route.userRoutes() {
         // Delete user
         delete {
             try {
-                val id = call.parameters["id"]?.toIntOrNull() ?: throw throw BadRequestException("Missing parameter")
-                dao.deleteUser(id)
-                call.respond(HttpStatusCode.OK)
+                val id = call.parameters["id"]?.toIntOrNull() ?: throw BadRequestException("Missing parameter")
+                val requesterId = call.principal<JWTPrincipal>()!!.payload.getClaim("userId").asString().toInt()// User ID in JWT Token
+
+                // Check if user is owner or admin belongs to the same org
+                if (requesterId == id|| isAdminSameOrg(requesterId, id)) {
+                    dao.deleteUser(id)
+                    call.respond(HttpStatusCode.OK)
+                } else {
+                    call.respond(HttpStatusCode.Forbidden, FORBIDDEN)
+                }
+
             } catch (e: BadRequestException) {
                 e.printStackTrace()
                 call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
+            } catch (e: NotFoundException) {
+                e.printStackTrace()
+                call.respond(HttpStatusCode.NotFound, NOT_FOUND)
             } catch (e: Exception) {
                 e.printStackTrace()
                 call.respond(HttpStatusCode.InternalServerError, FAILED_DELETE_USER)
@@ -78,6 +117,7 @@ fun Route.userRoutes() {
         }
     }
 
+    // Setup user account
     post("/user/{id}/setup") {
         try {
             val id = call.parameters["id"]?.toIntOrNull() ?: throw BadRequestException("Missing parameter")
@@ -101,6 +141,12 @@ fun Route.userRoutes() {
         } catch (e: BadRequestException) {
             e.printStackTrace()
             call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
+        } catch (e: ContentTransformationException) {
+            e.printStackTrace()
+            call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
+        } catch (e: IllegalArgumentException) {
+            e.printStackTrace()
+            call.respond(HttpStatusCode.BadRequest, BAD_REQUEST)
         } catch (e: NotFoundException) {
             e.printStackTrace()
             call.respond(HttpStatusCode.NotFound, NOT_FOUND)
@@ -109,4 +155,11 @@ fun Route.userRoutes() {
             call.respond(HttpStatusCode.InternalServerError, FAILED_SETUP)
         }
     }
+}
+
+// Check if the user is an admin in the same organization
+private suspend fun isAdminSameOrg(requesterId: Int, userId: Int): Boolean {
+    val requester = dao.getUser(requesterId)
+    val user = dao.getUser(userId) ?: throw NotFoundException()
+    return (requester?.admin == true) && (requester.orgId == user.orgId)
 }
